:lipstick: Refactor MainWindow menu bar into DesktopMenuBar component
